requirements.txt
pywin32
XInput-Python
pyautogui